Fix tests after router refactor
