+++ mypy.ini
@@ -32,6 +32,16 @@ ignore_errors = True
 ignore_errors = True
 [mypy-ume.api_deps]
 ignore_errors = True
+[mypy-ume.auth_routes]
+ignore_errors = True
+[mypy-ume.metrics_routes]
+ignore_errors = True
+[mypy-ume.dashboard_routes]
+ignore_errors = True
+[mypy-ume.pii_routes]
+ignore_errors = True
+[mypy-ume.recommendations_routes]
+ignore_errors = True
 
 [mypy-ume.config]
 ignore_errors = True
