+++ tests/test_api_consent.py
@@ -5,7 +5,7 @@ from ume.config import settings
 
 def _token(client: TestClient) -> str:
     res = client.post(
-        "/token",
+        "/auth/token",
         data={"username": settings.UME_OAUTH_USERNAME, "password": settings.UME_OAUTH_PASSWORD},
     )
     return res.json()["access_token"]
