+++ tests/test_metrics.py
@@ -62,7 +62,7 @@ def dummy_create() -> DummyVS:
     return DummyVS()
 
 package.VectorStore = DummyVS  # type: ignore[attr-defined]
-package.create_default_store = dummy_create  # type: ignore[attr-defined]
+package.create_vector_store = dummy_create  # type: ignore[attr-defined]
 spec_api = importlib.util.spec_from_file_location("ume.api", root / "src" / "ume" / "api.py")
 assert spec_api and spec_api.loader
 api_module = importlib.util.module_from_spec(spec_api)
@@ -126,7 +126,7 @@ def setup_module(_: object) -> None:
 
 def _token(client: TestClient) -> str:
     res = client.post(
-        "/token",
+        "/auth/token",
         data={
             "username": settings.UME_OAUTH_USERNAME,
             "password": settings.UME_OAUTH_PASSWORD,
